fix: remove `ember-cli-typescript`
I believe I need to move this addon to use `ember-auto-import`, and from what I can tell that is not compatible with TypeScript files (it does not see them as candidates to search for imports). As much as I would love to retain TypeeScript in this package, there are bugs to fix!